add SiPixelGainCalibrationForHLT_PayloadInspector, avoid duplication using templates, update tests and unit tests
