+++ CondCore/SiPixelPlugins/plugins/BuildFile.xml
@@ -21,3 +21,11 @@
   <use   name="CalibTracker/StandaloneTrackerTopology"/>
   <use   name="boost_python"/>
 </library>
+
+<library file="SiPixelGainCalibrationForHLT_PayloadInspector.cc" name="SiPixelGainCalibrationForHLT_PayloadInspector">
+  <use   name="CondCore/Utilities"/>
+  <use   name="CondCore/CondDB"/>
+  <use   name="CondFormats/Common"/>
+  <use   name="CalibTracker/StandaloneTrackerTopology"/>
+  <use   name="boost_python"/>
+</library>
