+++ CondCore/SiPixelPlugins/plugins/SiPixelGainCalibrationOffline_PayloadInspector.cc
@@ -1,1407 +1,83 @@
 /*!
   \file SiPixelGainCalibrationOffline_PayloadInspector
-  \Payload Inspector Plugin for SiPixel Gain Calibration for Offline
+  \Payload Inspector Plugin for SiPixel Gain Calibration Offline
   \author M. Musich
   \version $Revision: 1.0 $
   \date $Date: 2020/04/01 11:31:00 $
 */
 
-#include "CondCore/Utilities/interface/PayloadInspectorModule.h"
-#include "CondCore/Utilities/interface/PayloadInspector.h"
-#include "CondCore/CondDB/interface/Time.h"
+#include "CondCore/SiPixelPlugins/interface/SiPixelGainCalibHelper.h"
 #include "CondFormats/SiPixelObjects/interface/SiPixelGainCalibrationOffline.h"
-#include "FWCore/ParameterSet/interface/FileInPath.h"
-#include "CalibTracker/StandaloneTrackerTopology/interface/StandaloneTrackerTopology.h"
-#include "CondCore/SiPixelPlugins/interface/SiPixelPayloadInspectorHelper.h"
-#include "DataFormats/SiPixelDetId/interface/PixelSubdetector.h"
-
-#include <memory>
-#include <sstream>
-
-// include ROOT
-#include "TH2F.h"
-#include "TH1F.h"
-#include "TLegend.h"
-#include "TCanvas.h"
-#include "TLine.h"
-#include "TStyle.h"
-#include "TLatex.h"
-#include "TPave.h"
-#include "TPaveStats.h"
-#include "TGaxis.h"
 
 namespace {
 
-  namespace gainCalibPI {
-    enum type { t_gain = 0, t_pedestal = 1 };
-  }
-
-  constexpr char const* TypeName[2] = {"Gains", "Pedestals"};
-
-  using AvgMap = std::map<uint32_t, float>;
-
-  /*******************************************************************
-    1d histogram of SiPixelGainCalibrationOffline for Gains of 1 IOV 
-  ********************************************************************/
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineValues : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineValues()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Values", TypeName[myType])) {
-      setSingleIov(true);
-    }
-
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      gStyle->SetOptStat("emr");
-
-      auto iov = iovs.front();
-      std::shared_ptr<SiPixelGainCalibrationOffline> payload = fetchPayload(std::get<1>(iov));
-
-      float minimum(9999.);
-      float maximum(-9999.);
-
-      switch (myType) {
-        case gainCalibPI::t_gain:
-          maximum = payload->getGainHigh();
-          minimum = payload->getGainLow();
-          break;
-        case gainCalibPI::t_pedestal:
-          maximum = payload->getPedHigh();
-          minimum = payload->getPedLow();
-          break;
-        default:
-          edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Unrecognized type " << myType << std::endl;
-          break;
-      }
-
-      TCanvas canvas("Canv", "Canv", 1200, 1000);
-      auto h1 = std::make_shared<TH1F>(
-          Form("%s values", TypeName[myType]),
-          Form("SiPixel Gain Calibration Offline - %s;per pixel %s;# pixels", TypeName[myType], TypeName[myType]),
-          200,
-          minimum,
-          maximum);
-      canvas.cd();
-      canvas.SetTopMargin(0.06);
-      canvas.SetBottomMargin(0.12);
-      canvas.SetLeftMargin(0.12);
-      canvas.SetRightMargin(0.05);
-      canvas.Modified();
-
-      // fill the histogram
-      fillTheHisto(payload, h1, myType);
-
-      canvas.cd()->SetLogy();
-      h1->SetTitle("");
-      h1->GetYaxis()->SetRangeUser(0.1, h1->GetMaximum() * 10.);
-      h1->SetFillColor(kBlue);
-      h1->SetMarkerStyle(20);
-      h1->SetMarkerSize(1);
-      h1->Draw("bar2");
-
-      SiPixelPI::makeNicePlotStyle(h1.get());
-      h1->SetStats(true);
-
-      canvas.Update();
-
-      TLegend legend = TLegend(0.40, 0.88, 0.94, 0.93);
-      legend.SetHeader(("Payload hash: #bf{" + (std::get<1>(iov)) + "}").c_str(),
-                       "C");  // option "C" allows to center the header
-      //legend.AddEntry(h1.get(), ("IOV: " + std::to_string(std::get<0>(iov))).c_str(), "PL");
-      legend.SetLineColor(10);
-      legend.SetTextSize(0.025);
-      legend.Draw("same");
-
-      TPaveStats* st = (TPaveStats*)h1->FindObject("stats");
-      st->SetTextSize(0.03);
-      /*
-      st->SetX1NDC(0.70);  //new x start position
-      st->SetY1NDC(0.77);  //new y start position
-      st->SetX2NDC(0.94);  //new x end position
-      st->SetY2NDC(0.87);  //new y end position
-      */
-      st->SetX1NDC(0.15);  //new x start position
-      st->SetY1NDC(0.83);  //new y start position
-      st->SetX2NDC(0.39);  //new x end position
-      st->SetY2NDC(0.93);  //new y end position
-
-      auto ltx = TLatex();
-      ltx.SetTextFont(62);
-      //ltx.SetTextColor(kBlue);
-      ltx.SetTextSize(0.05);
-      ltx.SetTextAlign(11);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin() + 0.1,
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:" + std::to_string(std::get<0>(iov))).c_str());
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-
-      return true;
-    }
-
-    // method to avoid filling the histogram twice
-    void fillTheHisto(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                      std::shared_ptr<TH1F> h1,
-                      gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2 = 160;  // rows in x
-        //int ncols = 52 * 8 = 416;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float quid(-99999.);
-
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                quid = payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                quid = payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }
-
-            h1->Fill(quid);
-          }
-        }
-      }
-    }
-  };
-
-  typedef SiPixelGainCalibrationOfflineValues<gainCalibPI::t_gain> SiPixelGainCalibrationOfflineGainsValues;
-  typedef SiPixelGainCalibrationOfflineValues<gainCalibPI::t_pedestal> SiPixelGainCalibrationOfflinePedestalsValues;
-
-  /*******************************************************************
-    1d histogram of SiPixelGainCalibrationOffline for Gain/Pedestals
-    correlation of 1 IOV
-  ********************************************************************/
-  class SiPixelGainCalibrationCorrelations : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationCorrelations()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              "SiPixelGainCalibrationOffline gain/pedestal correlations") {
-      setSingleIov(true);
-    }
-
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      gStyle->SetOptStat("emr");
-      gStyle->SetPalette(1);
-
-      auto iov = iovs.front();
-      std::shared_ptr<SiPixelGainCalibrationOffline> payload = fetchPayload(std::get<1>(iov));
-
-      TCanvas canvas("Canv", "Canv", 1400, 800);
-      canvas.Divide(2, 1);
-      canvas.cd();
-
-      auto hBPix = std::make_shared<TH2F>("Correlation BPIX",
-                                          "SiPixel Gain Calibration Offline BPIx;per pixel gains;per pixel pedestals",
-                                          200,
-                                          payload->getGainLow(),
-                                          payload->getGainHigh(),
-                                          200,
-                                          payload->getPedLow(),
-                                          payload->getPedHigh());
-
-      auto hFPix = std::make_shared<TH2F>("Correlation FPIX",
-                                          "SiPixel Gain Calibration Offline FPix;per pixel gains;per pixel pedestals",
-                                          200,
-                                          payload->getGainLow(),
-                                          payload->getGainHigh(),
-                                          200,
-                                          payload->getPedLow(),
-                                          payload->getPedHigh());
-
-      for (unsigned int i : {1, 2}) {
-        canvas.cd(i)->SetTopMargin(0.04);
-        canvas.cd(i)->SetBottomMargin(0.12);
-        canvas.cd(i)->SetLeftMargin(0.15);
-        canvas.cd(i)->SetRightMargin(0.13);
-        canvas.cd(i)->Modified();
-      }
-
-      // actually fill the histograms
-      fillTheHistos(payload, hBPix, hFPix);
-
-      canvas.cd(1)->SetLogz();
-      hBPix->SetTitle("");
-      hBPix->Draw("colz");
-
-      SiPixelPI::makeNicePlotStyle(hBPix.get());
-      hBPix->GetYaxis()->SetTitleOffset(1.7);
-
-      canvas.cd(2)->SetLogz();
-      hFPix->SetTitle("");
-      hFPix->Draw("colz");
-
-      SiPixelPI::makeNicePlotStyle(hFPix.get());
-      hFPix->GetYaxis()->SetTitleOffset(1.7);
-      canvas.Update();
-
-      TLegend legend = TLegend(0.3, 0.92, 0.70, 0.95);
-      legend.SetHeader(("Payload hash: #bf{" + (std::get<1>(iov)) + "}").c_str(),
-                       "C");  // option "C" allows to center the header
-      legend.SetLineColor(10);
-      legend.SetTextSize(0.025);
-      canvas.cd(1);
-      legend.Draw("same");
-      canvas.cd(2);
-      legend.Draw("same");
-
-      auto ltx = TLatex();
-      ltx.SetTextFont(62);
-      //ltx.SetTextColor(kBlue);
-      ltx.SetTextSize(0.045);
-      ltx.SetTextAlign(11);
-      canvas.cd(1);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin() + 0.01,
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:" + std::to_string(std::get<0>(iov))).c_str());
-
-      ltx.DrawLatexNDC(0.75, 0.15, "BPIX");
-
-      canvas.cd(2);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin() + 0.01,
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:" + std::to_string(std::get<0>(iov))).c_str());
-
-      ltx.DrawLatexNDC(0.75, 0.15, "FPIX");
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-#ifdef MMDEBUG
-      canvas.SaveAs("out.root");
-#endif
-      return true;
-    }
-
-    // method to avoid filling the histogram twice
-    void fillTheHistos(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                       std::shared_ptr<TH2F> hBPix,
-                       std::shared_ptr<TH2F> hFPix) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        int subid = DetId(d).subdetId();
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            float gain = payload->getGain(
-                col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-            float ped =
-                payload->getPed(col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-
-            switch (subid) {
-              case PixelSubdetector::PixelBarrel:
-                hBPix->Fill(gain, ped);
-                break;
-              case PixelSubdetector::PixelEndcap:
-                hFPix->Fill(gain, ped);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << d << " is not a Pixel DetId" << std::endl;
-                break;
-            }
-          }
-        }
-      }
-    }
-  };
-
-  /*******************************************************************
-    1d histogram of SiPixelGainCalibrationOffline for Pedestals of 1 IOV
-  ********************************************************************/
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineValuesByPart
-      : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineValuesByPart()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Values By Partition", TypeName[myType])) {
-      setSingleIov(true);
-    }
-
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      gStyle->SetOptStat("emr");
-
-      auto iov = iovs.front();
-      std::shared_ptr<SiPixelGainCalibrationOffline> payload = fetchPayload(std::get<1>(iov));
-
-      TCanvas canvas("Canv", "Canv", 1400, 800);
-      canvas.Divide(2, 1);
-      canvas.cd();
-
-      float minimum(9999.);
-      float maximum(-9999.);
-
-      switch (myType) {
-        case gainCalibPI::t_gain:
-          maximum = payload->getGainHigh();
-          minimum = payload->getGainLow();
-          break;
-        case gainCalibPI::t_pedestal:
-          maximum = payload->getPedHigh();
-          minimum = payload->getPedLow();
-          break;
-        default:
-          edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Unrecognized type " << myType << std::endl;
-          break;
-      }
-
-      auto hBPix = std::make_shared<TH1F>(Form("%s BPIX", TypeName[myType]),
-                                          Form("SiPixel Gain Calibration Offline BPIx -%s;per pixel %s (BPix);# pixels",
-                                               TypeName[myType],
-                                               TypeName[myType]),
-                                          200,
-                                          minimum,
-                                          maximum);
-
-      auto hFPix = std::make_shared<TH1F>(Form("%s FPIX", TypeName[myType]),
-                                          Form("SiPixel Gain Calibration Offline FPix -%s;per pixel %s (FPix);# pixels",
-                                               TypeName[myType],
-                                               TypeName[myType]),
-                                          200,
-                                          minimum,
-                                          maximum);
-
-      for (unsigned int i : {1, 2}) {
-        canvas.cd(i)->SetTopMargin(0.04);
-        canvas.cd(i)->SetBottomMargin(0.12);
-        canvas.cd(i)->SetLeftMargin(0.12);
-        canvas.cd(i)->SetRightMargin(0.02);
-        canvas.cd(i)->Modified();
-      }
-
-      // actually fill the histograms
-      fillTheHistos(payload, hBPix, hFPix, myType);
-
-      canvas.cd(1)->SetLogy();
-      hBPix->SetTitle("");
-      hBPix->GetYaxis()->SetRangeUser(0.1, hBPix->GetMaximum() * 10);
-      hBPix->SetFillColor(kBlue);
-      hBPix->SetMarkerStyle(20);
-      hBPix->SetMarkerSize(1);
-      //hBPix->Draw("bar2");
-      hBPix->Draw("hist");
-
-      SiPixelPI::makeNicePlotStyle(hBPix.get());
-      hBPix->SetStats(true);
-
-      canvas.cd(2)->SetLogy();
-      hFPix->SetTitle("");
-      hFPix->GetYaxis()->SetRangeUser(0.1, hFPix->GetMaximum() * 10);
-      hFPix->SetFillColor(kBlue);
-      hFPix->SetMarkerStyle(20);
-      hFPix->SetMarkerSize(1);
-      //hFPix->Draw("bar2");
-      hFPix->Draw("hist");
-
-      SiPixelPI::makeNicePlotStyle(hFPix.get());
-      hFPix->SetStats(true);
-
-      canvas.Update();
-
-      TLegend legend = TLegend(0.32, 0.92, 0.97, 0.95);
-      legend.SetHeader(("Payload hash: #bf{" + (std::get<1>(iov)) + "}").c_str(),
-                       "C");  // option "C" allows to center the header
-      //legend.AddEntry(h1.get(), ("IOV: " + std::to_string(std::get<0>(iov))).c_str(), "PL");
-      legend.SetLineColor(10);
-      legend.SetTextSize(0.025);
-      canvas.cd(1);
-      legend.Draw("same");
-      canvas.cd(2);
-      legend.Draw("same");
-
-      canvas.cd(1);
-      TPaveStats* st1 = (TPaveStats*)hBPix->FindObject("stats");
-      st1->SetTextSize(0.03);
-      st1->SetX1NDC(0.13);   //new x start position
-      st1->SetY1NDC(0.815);  //new y start position
-      st1->SetX2NDC(0.44);   //new x end position
-      st1->SetY2NDC(0.915);  //new y end position
-
-      canvas.cd(2);
-      TPaveStats* st2 = (TPaveStats*)hFPix->FindObject("stats");
-      st2->SetTextSize(0.03);
-      st2->SetX1NDC(0.14);   //new x start position
-      st2->SetY1NDC(0.815);  //new y start position
-      st2->SetX2NDC(0.44);   //new x end position
-      st2->SetY2NDC(0.915);  //new y end position
-
-      auto ltx = TLatex();
-      ltx.SetTextFont(62);
-      //ltx.SetTextColor(kBlue);
-      ltx.SetTextSize(0.045);
-      ltx.SetTextAlign(11);
-      canvas.cd(1);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin() + 0.01,
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:" + std::to_string(std::get<0>(iov))).c_str());
-
-      canvas.cd(2);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin() + 0.01,
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:" + std::to_string(std::get<0>(iov))).c_str());
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-
-      return true;
-    }
-
-    // method to avoid filling the histogram twice
-    void fillTheHistos(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                       std::shared_ptr<TH1F> hBPix,
-                       std::shared_ptr<TH1F> hFPix,
-                       gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        int subid = DetId(d).subdetId();
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2;  // rows in x
-        //int ncols = 52 * 8;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float quid(-99999.);
-
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                quid = payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                quid = payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }
-
-            switch (subid) {
-              case PixelSubdetector::PixelBarrel:
-                hBPix->Fill(quid);
-                break;
-              case PixelSubdetector::PixelEndcap:
-                hFPix->Fill(quid);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << d << " is not a Pixel DetId" << std::endl;
-                break;
-            }
-          }
-        }
-      }
-    }
-  };
-
-  typedef SiPixelGainCalibrationOfflineValuesByPart<gainCalibPI::t_gain> SiPixelGainCalibrationOfflineGainsByPart;
-  typedef SiPixelGainCalibrationOfflineValuesByPart<gainCalibPI::t_pedestal>
-      SiPixelGainCalibrationOfflinePedestalsByPart;
-
-  /************************************************
-    1d histogram comparison of SiPixelGainCalibrationOffline
-  *************************************************/
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineValueComparisonBase
-      : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineValueComparisonBase()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Values Comparison", TypeName[myType])) {}
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      gStyle->SetOptStat("emr");
-      TGaxis::SetExponentOffset(-0.1, 0.01, "y");  // Y offset
-      TH1F::SetDefaultSumw2(true);
-
-      std::vector<std::tuple<cond::Time_t, cond::Hash>> sorted_iovs = iovs;
-      // make absolute sure the IOVs are sortd by since
-      std::sort(begin(sorted_iovs), end(sorted_iovs), [](auto const& t1, auto const& t2) {
-        return std::get<0>(t1) < std::get<0>(t2);
-      });
-      auto firstiov = sorted_iovs.front();
-      auto lastiov = sorted_iovs.back();
-
-      std::shared_ptr<SiPixelGainCalibrationOffline> last_payload = fetchPayload(std::get<1>(lastiov));
-      std::shared_ptr<SiPixelGainCalibrationOffline> first_payload = fetchPayload(std::get<1>(firstiov));
-
-      std::string lastIOVsince = std::to_string(std::get<0>(lastiov));
-      std::string firstIOVsince = std::to_string(std::get<0>(firstiov));
-
-      float minimum(9999.);
-      float maximum(-9999.);
-
-      switch (myType) {
-        case gainCalibPI::t_gain:
-          maximum = std::max(last_payload->getGainHigh(), first_payload->getGainHigh());
-          minimum = std::min(last_payload->getGainLow(), first_payload->getGainLow());
-          break;
-        case gainCalibPI::t_pedestal:
-          maximum = std::max(last_payload->getPedHigh(), first_payload->getPedHigh());
-          minimum = std::min(last_payload->getPedLow(), first_payload->getPedLow());
-          break;
-        default:
-          edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Unrecognized type " << myType << std::endl;
-          break;
-      }
-
-      TCanvas canvas("Canv", "Canv", 1200, 1000);
-      canvas.cd();
-      auto hfirst = std::make_shared<TH1F>(
-          Form("First, IOV %s", firstIOVsince.c_str()),
-          Form("SiPixel Gain Calibration Offline - %s;per pixel %s;# pixels", TypeName[myType], TypeName[myType]),
-          200,
-          minimum,
-          maximum);
-
-      auto hlast = std::make_shared<TH1F>(
-          Form("Last, IOV %s", lastIOVsince.c_str()),
-          Form("SiPixel Gain Calibration Offline - %s;per pixel %s;# pixels", TypeName[myType], TypeName[myType]),
-          200,
-          minimum,
-          maximum);
-
-      canvas.SetTopMargin(0.05);
-      canvas.SetBottomMargin(0.12);
-      canvas.SetLeftMargin(0.12);
-      canvas.SetRightMargin(0.03);
-      canvas.Modified();
-
-      fillTheHisto(first_payload, hfirst, myType);
-      fillTheHisto(last_payload, hlast, myType);
-
-      canvas.cd()->SetLogy();
-      auto extrema = SiPixelPI::getExtrema(hfirst.get(), hlast.get());
-      //hfirst->GetYaxis()->SetRangeUser(extrema.first, extrema.second * 1.10);
-      hfirst->GetYaxis()->SetRangeUser(1., extrema.second * 10);
-
-      hfirst->SetTitle("");
-      //hfirst->SetFillColor(kRed);
-      hfirst->SetLineColor(kRed);
-      hfirst->SetBarWidth(0.95);
-      //hfirst->Draw("histbar");
-      hfirst->Draw("hist");
-
-      hlast->SetTitle("");
-      hlast->SetFillColorAlpha(kBlue, 0.20);
-      hlast->SetBarWidth(0.95);
-      //hlast->Draw("histbarsame");
-      hlast->Draw("histsames");
-
-      SiPixelPI::makeNicePlotStyle(hfirst.get());
-      hfirst->SetStats(true);
-      SiPixelPI::makeNicePlotStyle(hlast.get());
-      hlast->SetStats(true);
-
-      canvas.Update();
-
-      TLegend legend = TLegend(0.45, 0.86, 0.74, 0.94);
-      //legend.SetHeader("#font[22]{SiPixel Offline Gain Calibration Comparison}", "C");  // option "C" allows to center the header
-      //legend.AddEntry(hfirst.get(), ("IOV: " + std::to_string(std::get<0>(firstiov))).c_str(), "FL");
-      //legend.AddEntry(hlast.get(),  ("IOV: " + std::to_string(std::get<0>(lastiov))).c_str(), "FL");
-      legend.AddEntry(hfirst.get(), ("payload: #color[2]{" + std::get<1>(firstiov) + "}").c_str(), "F");
-      legend.AddEntry(hlast.get(), ("payload: #color[4]{" + std::get<1>(lastiov) + "}").c_str(), "F");
-      legend.SetTextSize(0.022);
-      legend.SetLineColor(10);
-      legend.Draw("same");
-
-      TPaveStats* st1 = (TPaveStats*)hfirst->FindObject("stats");
-      st1->SetTextSize(0.022);
-      st1->SetLineColor(kRed);
-      st1->SetTextColor(kRed);
-      st1->SetX1NDC(0.13);  //new x start position
-      st1->SetY1NDC(0.84);  //new y start position
-      st1->SetX2NDC(0.31);  //new x end position
-      st1->SetY2NDC(0.94);  //new y end position
-
-      TPaveStats* st2 = (TPaveStats*)hlast->FindObject("stats");
-      st2->SetTextSize(0.022);
-      st2->SetLineColor(kBlue);
-      st2->SetTextColor(kBlue);
-      st2->SetX1NDC(0.13);  //new x start position
-      st2->SetY1NDC(0.73);  //new y start position
-      st2->SetX2NDC(0.31);  //new x end position
-      st2->SetY2NDC(0.83);  //new y end position
-
-      auto ltx = TLatex();
-      ltx.SetTextFont(62);
-      //ltx.SetTextColor(kBlue);
-      ltx.SetTextSize(0.047);
-      ltx.SetTextAlign(11);
-      ltx.DrawLatexNDC(gPad->GetLeftMargin(),
-                       1 - gPad->GetTopMargin() + 0.01,
-                       ("SiPixel Gain Calibration IOV:#color[2]{" + std::to_string(std::get<0>(firstiov)) +
-                        "} vs IOV:#color[4]{" + std::to_string(std::get<0>(lastiov)) + "}")
-                           .c_str());
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-#ifdef MMDEBUG
-      canvas.SaveAs("out.root");
-#endif
-
-      return true;
-    }
-
-    // method to avoid filling the histogram twice
-    void fillTheHisto(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                      std::shared_ptr<TH1F> h1,
-                      gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2;  // rows in x
-        //int ncols = 52 * 8;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float quid(-99999.);
-
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                quid = payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                quid = payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }
-            h1->Fill(quid);
-          }
-        }
-      }
-    }
-  };
-
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineValueComparisonSingleTag
-      : public SiPixelGainCalibrationOfflineValueComparisonBase<myType> {
-  public:
-    SiPixelGainCalibrationOfflineValueComparisonSingleTag()
-        : SiPixelGainCalibrationOfflineValueComparisonBase<myType>() {
-      this->setSingleIov(false);
-    }
-  };
-
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineValueComparisonTwoTags
-      : public SiPixelGainCalibrationOfflineValueComparisonBase<myType> {
-  public:
-    SiPixelGainCalibrationOfflineValueComparisonTwoTags() : SiPixelGainCalibrationOfflineValueComparisonBase<myType>() {
-      this->setTwoTags(true);
-    }
-  };
-
-  typedef SiPixelGainCalibrationOfflineValueComparisonSingleTag<gainCalibPI::t_gain>
-      SiPixelGainCalibOfflineGainComparisonSingleTag;
-  typedef SiPixelGainCalibrationOfflineValueComparisonSingleTag<gainCalibPI::t_pedestal>
-      SiPixelGainCalibOfflinePedestalComparisonSingleTag;
-
-  typedef SiPixelGainCalibrationOfflineValueComparisonTwoTags<gainCalibPI::t_gain>
-      SiPixelGainCalibOfflineGainComparisonTwoTags;
-  typedef SiPixelGainCalibrationOfflineValueComparisonTwoTags<gainCalibPI::t_pedestal>
-      SiPixelGainCalibOfflinePedestalComparisonTwoTags;
-
-  // 2D MAPS
-
-  /************************************************
-   occupancy style map BPix
-  *************************************************/
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineBPIXMap : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineBPIXMap()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Barrel Pixel Map", TypeName[myType])),
-          m_trackerTopo{StandaloneTrackerTopology::fromTrackerParametersXMLFile(
-              edm::FileInPath("Geometry/TrackerCommonData/data/PhaseI/trackerParameters.xml").fullPath())} {
-      setSingleIov(true);
-    };
-
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      auto iov = iovs.front();
-      std::shared_ptr<SiPixelGainCalibrationOffline> payload = fetchPayload(std::get<1>(iov));
-
-      static const int n_layers = 4;
-      int nlad_list[n_layers] = {6, 14, 22, 32};
-      int divide_roc = 1;
-
-      // ---------------------    BOOK HISTOGRAMS
-      std::array<TH2D*, n_layers> h_bpix_Gains;
-
-      for (unsigned int lay = 1; lay <= 4; lay++) {
-        int nlad = nlad_list[lay - 1];
-
-        std::string name = "occ_Gain_Layer_" + std::to_string(lay);
-        std::string title = "; Module # ; Ladder #";
-        h_bpix_Gains[lay - 1] = new TH2D(name.c_str(),
-                                         title.c_str(),
-                                         72 * divide_roc,
-                                         -4.5,
-                                         4.5,
-                                         (nlad * 4 + 2) * divide_roc,
-                                         -nlad - 0.5,
-                                         nlad + 0.5);
-      }
-
-      std::map<uint32_t, float> GainCalibMap_;
-      fillThePerModuleMap(payload, GainCalibMap_, myType);
-
-      // hard-coded phase-I
-      std::array<double, 4> minima = {{999., 999., 999., 999.}};
-
-      for (const auto& element : GainCalibMap_) {
-        int subid = DetId(element.first).subdetId();
-        if (subid == PixelSubdetector::PixelBarrel) {
-          auto layer = m_trackerTopo.pxbLayer(DetId(element.first));
-          auto s_ladder = SiPixelPI::signed_ladder(DetId(element.first), m_trackerTopo, true);
-          auto s_module = SiPixelPI::signed_module(DetId(element.first), m_trackerTopo, true);
-
-          auto ladder = m_trackerTopo.pxbLadder(DetId(element.first));
-          auto module = m_trackerTopo.pxbModule(DetId(element.first));
-          COUT << "layer:" << layer << " ladder:" << ladder << " module:" << module << " signed ladder: " << s_ladder
-               << " signed module: " << s_module << std::endl;
-
-          if (element.second < minima.at(layer - 1))
-            minima.at(layer - 1) = element.second;
-
-          auto rocsToMask = SiPixelPI::maskedBarrelRocsToBins(layer, s_ladder, s_module);
-          for (const auto& bin : rocsToMask) {
-            h_bpix_Gains[layer - 1]->SetBinContent(bin.first, bin.second, element.second);
-          }
-        }
-      }
-
-      gStyle->SetOptStat(0);
-      //=========================
-      TCanvas canvas("Summary", "Summary", 1200, 1200);
-      canvas.Divide(2, 2);
-
-      for (unsigned int lay = 1; lay <= 4; lay++) {
-        canvas.cd(lay)->SetBottomMargin(0.08);
-        canvas.cd(lay)->SetLeftMargin(0.1);
-        canvas.cd(lay)->SetRightMargin(0.13);
-
-        COUT << " layer:" << lay << " max:" << h_bpix_Gains[lay - 1]->GetMaximum() << " min: " << minima.at(lay - 1)
-             << std::endl;
-
-        SiPixelPI::dress_occup_plot(canvas, h_bpix_Gains[lay - 1], lay, 0, 1, true, true, false);
-        h_bpix_Gains[lay - 1]->GetZaxis()->SetRangeUser(minima.at(lay - 1) - 0.001,
-                                                        h_bpix_Gains[lay - 1]->GetMaximum() + 0.001);
-      }
-
-      auto unpacked = SiPixelPI::unpack(std::get<0>(iov));
-
-      for (unsigned int lay = 1; lay <= 4; lay++) {
-        canvas.cd(lay);
-        auto ltx = TLatex();
-        ltx.SetTextFont(62);
-        ltx.SetTextColor(kBlue);
-        ltx.SetTextSize(0.055);
-        ltx.SetTextAlign(11);
-        ltx.DrawLatexNDC(gPad->GetLeftMargin(),
-                         1 - gPad->GetTopMargin() + 0.01,
-                         unpacked.first == 0
-                             ? ("IOV:" + std::to_string(unpacked.second)).c_str()
-                             : (std::to_string(unpacked.first) + "," + std::to_string(unpacked.second)).c_str());
-      }
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-
-      return true;
-    }
-
-    void fillThePerModuleMap(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                             AvgMap& map,
-                             gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2;  // rows in x
-        //int ncols = 52 * 8;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float sumOfX(0.);
-        int nPixels(0);
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            nPixels++;
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                sumOfX += payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                sumOfX += payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }  // switch on the type
-          }    // rows
-        }      // columns
-        // fill the return value map
-        map[d] = sumOfX / nPixels;
-      }  // loop on the detId
-    }
-
-  private:
-    TrackerTopology m_trackerTopo;
-  };
-
-  /************************************************
-   occupancy style map FPix
-  *************************************************/
-
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineFPIXMap : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineFPIXMap()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Forward Pixel Map", TypeName[myType])),
-          m_trackerTopo{StandaloneTrackerTopology::fromTrackerParametersXMLFile(
-              edm::FileInPath("Geometry/TrackerCommonData/data/PhaseI/trackerParameters.xml").fullPath())} {
-      setSingleIov(true);
-    }
-
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      auto iov = iovs.front();
-      std::shared_ptr<SiPixelGainCalibrationOffline> payload = fetchPayload(std::get<1>(iov));
-
-      static const int n_rings = 2;
-      std::array<TH2D*, n_rings> h_fpix_Gains;
-      int divide_roc = 1;
-
-      // ---------------------    BOOK HISTOGRAMS
-      for (unsigned int ring = 1; ring <= n_rings; ring++) {
-        int n = ring == 1 ? 92 : 140;
-        float y = ring == 1 ? 11.5 : 17.5;
-        std::string name = "occ_Gains_ring_" + std::to_string(ring);
-        std::string title = "; Disk # ; Blade/Panel #";
-
-        h_fpix_Gains[ring - 1] =
-            new TH2D(name.c_str(), title.c_str(), 56 * divide_roc, -3.5, 3.5, n * divide_roc, -y, y);
-      }
-
-      std::map<uint32_t, float> GainCalibMap_;
-      fillThePerModuleMap(payload, GainCalibMap_, myType);
-
-      // hardcoded phase-I
-      std::array<double, 2> minima = {{999., 999.}};
-
-      for (const auto& element : GainCalibMap_) {
-        int subid = DetId(element.first).subdetId();
-        if (subid == PixelSubdetector::PixelEndcap) {
-          auto ring = SiPixelPI::ring(DetId(element.first), m_trackerTopo, true);
-          auto s_blade = SiPixelPI::signed_blade(DetId(element.first), m_trackerTopo, true);
-          auto s_disk = SiPixelPI::signed_disk(DetId(element.first), m_trackerTopo, true);
-          auto s_blade_panel = SiPixelPI::signed_blade_panel(DetId(element.first), m_trackerTopo, true);
-          auto panel = m_trackerTopo.pxfPanel(element.first);
-
-          COUT << "ring:" << ring << " blade: " << s_blade << " panel: " << panel
-               << " signed blade/panel: " << s_blade_panel << " disk: " << s_disk << std::endl;
-
-          if (element.second < minima.at(ring - 1))
-            minima.at(ring - 1) = element.second;
-
-          auto rocsToMask = SiPixelPI::maskedForwardRocsToBins(ring, s_blade, panel, s_disk);
-          for (const auto& bin : rocsToMask) {
-            h_fpix_Gains[ring - 1]->SetBinContent(bin.first, bin.second, element.second);
-          }
-        }
-      }
-
-      gStyle->SetOptStat(0);
-      //=========================
-      TCanvas canvas("Summary", "Summary", 1200, 600);
-      canvas.Divide(2, 1);
-
-      for (unsigned int ring = 1; ring <= n_rings; ring++) {
-        canvas.cd(ring)->SetBottomMargin(0.08);
-        canvas.cd(ring)->SetLeftMargin(0.1);
-        canvas.cd(ring)->SetRightMargin(0.13);
-
-        COUT << " ringer:" << ring << " max:" << h_fpix_Gains[ring - 1]->GetMaximum() << " min: " << minima.at(ring - 1)
-             << std::endl;
-
-        SiPixelPI::dress_occup_plot(canvas, h_fpix_Gains[ring - 1], 0, ring, 1, true, true, false);
-        h_fpix_Gains[ring - 1]->GetZaxis()->SetRangeUser(minima.at(ring - 1) - 0.001,
-                                                         h_fpix_Gains[ring - 1]->GetMaximum() + 0.001);
-      }
-
-      auto unpacked = SiPixelPI::unpack(std::get<0>(iov));
-
-      for (unsigned int ring = 1; ring <= n_rings; ring++) {
-        canvas.cd(ring);
-        auto ltx = TLatex();
-        ltx.SetTextFont(62);
-        ltx.SetTextColor(kBlue);
-        ltx.SetTextSize(0.05);
-        ltx.SetTextAlign(11);
-        ltx.DrawLatexNDC(gPad->GetLeftMargin(),
-                         1 - gPad->GetTopMargin() + 0.01,
-                         unpacked.first == 0
-                             ? ("IOV:" + std::to_string(unpacked.second)).c_str()
-                             : (std::to_string(unpacked.first) + "," + std::to_string(unpacked.second)).c_str());
-      }
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-      return true;
-    }
-
-    void fillThePerModuleMap(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                             AvgMap& map,
-                             gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2;  // rows in x
-        //int ncols = 52 * 8;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float sumOfX(0.);
-        int nPixels(0);
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            nPixels++;
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                sumOfX += payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                sumOfX += payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }  // switch on the type
-          }    // rows
-        }      // columns
-        // fill the return value map
-        map[d] = sumOfX / nPixels;
-      }  // loop on the detId
-    }
-
-  private:
-    TrackerTopology m_trackerTopo;
-  };
-
-  typedef SiPixelGainCalibrationOfflineBPIXMap<gainCalibPI::t_gain> SiPixelGainCalibOfflineGainsBPIXMap;
-  typedef SiPixelGainCalibrationOfflineBPIXMap<gainCalibPI::t_pedestal> SiPixelGainCalibOfflinePedestalsBPIXMap;
-
-  typedef SiPixelGainCalibrationOfflineFPIXMap<gainCalibPI::t_gain> SiPixelGainCalibOfflineGainsFPIXMap;
-  typedef SiPixelGainCalibrationOfflineFPIXMap<gainCalibPI::t_pedestal> SiPixelGainCalibOfflinePedestalsFPIXMap;
-
-  /************************************************
-   Summary Comparison per region of SiPixelGainCalibrationOffline between 2 IOVs
-  *************************************************/
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineByRegionComparisonBase
-      : public cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline> {
-  public:
-    SiPixelGainCalibrationOfflineByRegionComparisonBase()
-        : cond::payloadInspector::PlotImage<SiPixelGainCalibrationOffline>(
-              Form("SiPixelGainCalibrationOffline %s Comparison by Region", TypeName[myType])) {}
-    bool fill(const std::vector<std::tuple<cond::Time_t, cond::Hash>>& iovs) override {
-      gStyle->SetPaintTextFormat(".3f");
-
-      std::vector<std::tuple<cond::Time_t, cond::Hash>> sorted_iovs = iovs;
-
-      // make absolute sure the IOVs are sortd by since
-      std::sort(begin(sorted_iovs), end(sorted_iovs), [](auto const& t1, auto const& t2) {
-        return std::get<0>(t1) < std::get<0>(t2);
-      });
-
-      auto firstiov = sorted_iovs.front();
-      auto lastiov = sorted_iovs.back();
-
-      std::shared_ptr<SiPixelGainCalibrationOffline> last_payload = fetchPayload(std::get<1>(lastiov));
-      std::shared_ptr<SiPixelGainCalibrationOffline> first_payload = fetchPayload(std::get<1>(firstiov));
-
-      std::map<uint32_t, float> f_GainsMap_;
-      fillThePerModuleMap(first_payload, f_GainsMap_, myType);
-
-      std::map<uint32_t, float> l_GainsMap_;
-      fillThePerModuleMap(last_payload, l_GainsMap_, myType);
-
-      std::string lastIOVsince = std::to_string(std::get<0>(lastiov));
-      std::string firstIOVsince = std::to_string(std::get<0>(firstiov));
-
-      TCanvas canvas("Comparison", "Comparison", 1600, 800);
-
-      std::map<SiPixelPI::regions, std::shared_ptr<TH1F>> FirstGains_spectraByRegion;
-      std::map<SiPixelPI::regions, std::shared_ptr<TH1F>> LastGains_spectraByRegion;
-      std::shared_ptr<TH1F> summaryFirst;
-      std::shared_ptr<TH1F> summaryLast;
-
-      float minimum(9999.);
-      float maximum(-9999.);
-
-      switch (myType) {
-        case gainCalibPI::t_gain:
-          maximum = std::max(last_payload->getGainHigh(), first_payload->getGainHigh());
-          minimum = std::min(last_payload->getGainLow(), first_payload->getGainLow());
-          break;
-        case gainCalibPI::t_pedestal:
-          maximum = std::max(last_payload->getPedHigh(), first_payload->getPedHigh());
-          minimum = std::min(last_payload->getPedLow(), first_payload->getPedLow());
-          break;
-        default:
-          edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Unrecognized type " << myType << std::endl;
-          break;
-      }
-
-      // book the intermediate histograms
-      for (int r = SiPixelPI::BPixL1o; r != SiPixelPI::NUM_OF_REGIONS; r++) {
-        SiPixelPI::regions part = static_cast<SiPixelPI::regions>(r);
-        std::string s_part = SiPixelPI::getStringFromRegionEnum(part);
-
-        FirstGains_spectraByRegion[part] =
-            std::make_shared<TH1F>(Form("hfirstGains_%s", s_part.c_str()),
-                                   Form(";%s #LT %s #GT;n. of modules", s_part.c_str(), TypeName[myType]),
-                                   200,
-                                   minimum,
-                                   maximum);
-
-        LastGains_spectraByRegion[part] =
-            std::make_shared<TH1F>(Form("hlastGains_%s", s_part.c_str()),
-                                   Form(";%s #LT %s #GT;n. of modules", s_part.c_str(), TypeName[myType]),
-                                   200,
-                                   minimum,
-                                   maximum);
-      }
-
-      summaryFirst = std::make_shared<TH1F>(
-          "first Summary",
-          Form("Summary of #LT per pixel %s #GT;;average %s", TypeName[myType], TypeName[myType]),
-          FirstGains_spectraByRegion.size(),
-          0,
-          FirstGains_spectraByRegion.size());
-
-      summaryLast = std::make_shared<TH1F>(
-          "last Summary",
-          Form("Summary of #LT per pixel %s #GT;;average %s", TypeName[myType], TypeName[myType]),
-          LastGains_spectraByRegion.size(),
-          0,
-          LastGains_spectraByRegion.size());
-
-      const char* path_toTopologyXML = (f_GainsMap_.size() == SiPixelPI::phase0size)
-                                           ? "Geometry/TrackerCommonData/data/trackerParameters.xml"
-                                           : "Geometry/TrackerCommonData/data/PhaseI/trackerParameters.xml";
-      TrackerTopology f_tTopo =
-          StandaloneTrackerTopology::fromTrackerParametersXMLFile(edm::FileInPath(path_toTopologyXML).fullPath());
-
-      bool isPhase0(false);
-      if (f_GainsMap_.size() == SiPixelPI::phase0size) {
-        isPhase0 = true;
-      }
-
-      // -------------------------------------------------------------------
-      // loop on the first Gains Map
-      // -------------------------------------------------------------------
-      for (const auto& it : f_GainsMap_) {
-        if (DetId(it.first).det() != DetId::Tracker) {
-          edm::LogWarning("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Encountered invalid Tracker DetId:" << it.first << " - terminating ";
-          return false;
-        }
-
-        SiPixelPI::topolInfo t_info_fromXML;
-        t_info_fromXML.init();
-        DetId detid(it.first);
-        t_info_fromXML.fillGeometryInfo(detid, f_tTopo, isPhase0);
-
-        SiPixelPI::regions thePart = t_info_fromXML.filterThePartition();
-        FirstGains_spectraByRegion[thePart]->Fill(it.second);
-      }  // ends loop on the vector of error transforms
-
-      path_toTopologyXML = (l_GainsMap_.size() == SiPixelPI::phase0size)
-                               ? "Geometry/TrackerCommonData/data/trackerParameters.xml"
-                               : "Geometry/TrackerCommonData/data/PhaseI/trackerParameters.xml";
-      TrackerTopology l_tTopo =
-          StandaloneTrackerTopology::fromTrackerParametersXMLFile(edm::FileInPath(path_toTopologyXML).fullPath());
-
-      if (l_GainsMap_.size() == SiPixelPI::phase0size) {
-        isPhase0 = true;
-      }
-
-      // -------------------------------------------------------------------
-      // loop on the second Gains Map
-      // -------------------------------------------------------------------
-      for (const auto& it : l_GainsMap_) {
-        if (DetId(it.first).det() != DetId::Tracker) {
-          edm::LogWarning("SiPixelGainCalibrationOffline_PayloadInspector")
-              << "Encountered invalid Tracker DetId:" << it.first << " - terminating ";
-          return false;
-        }
-
-        SiPixelPI::topolInfo t_info_fromXML;
-        t_info_fromXML.init();
-        DetId detid(it.first);
-        t_info_fromXML.fillGeometryInfo(detid, l_tTopo, isPhase0);
-
-        SiPixelPI::regions thePart = t_info_fromXML.filterThePartition();
-        LastGains_spectraByRegion[thePart]->Fill(it.second);
-      }  // ends loop on the vector of error transforms
-
-      // fill the summary plots
-      int bin = 1;
-      for (int r = SiPixelPI::BPixL1o; r != SiPixelPI::NUM_OF_REGIONS; r++) {
-        SiPixelPI::regions part = static_cast<SiPixelPI::regions>(r);
-
-        summaryFirst->GetXaxis()->SetBinLabel(bin, SiPixelPI::getStringFromRegionEnum(part).c_str());
-        // avoid filling the histogram with numerical noise
-        float f_mean =
-            FirstGains_spectraByRegion[part]->GetMean() > 10.e-6 ? FirstGains_spectraByRegion[part]->GetMean() : 0.;
-        summaryFirst->SetBinContent(bin, f_mean);
-        //summaryFirst->SetBinError(bin,Gains_spectraByRegion[hash]->GetRMS());
-
-        summaryLast->GetXaxis()->SetBinLabel(bin, SiPixelPI::getStringFromRegionEnum(part).c_str());
-        // avoid filling the histogram with numerical noise
-        float l_mean =
-            LastGains_spectraByRegion[part]->GetMean() > 10.e-6 ? LastGains_spectraByRegion[part]->GetMean() : 0.;
-        summaryLast->SetBinContent(bin, l_mean);
-        //summaryLast->SetBinError(bin,Gains_spectraByRegion[hash]->GetRMS());
-        bin++;
-      }
-
-      SiPixelPI::makeNicePlotStyle(summaryFirst.get());  //, kBlue);
-      summaryFirst->SetMarkerColor(kRed);
-      summaryFirst->GetXaxis()->LabelsOption("v");
-      summaryFirst->GetXaxis()->SetLabelSize(0.05);
-      summaryFirst->GetYaxis()->SetTitleOffset(0.9);
-
-      SiPixelPI::makeNicePlotStyle(summaryLast.get());  //, kRed);
-      summaryLast->SetMarkerColor(kBlue);
-      summaryLast->GetYaxis()->SetTitleOffset(0.9);
-      summaryLast->GetXaxis()->LabelsOption("v");
-      summaryLast->GetXaxis()->SetLabelSize(0.05);
-
-      canvas.cd()->SetGridy();
-
-      canvas.SetBottomMargin(0.18);
-      canvas.SetLeftMargin(0.11);
-      canvas.SetRightMargin(0.02);
-      canvas.Modified();
-
-      summaryFirst->SetFillColor(kRed);
-      summaryLast->SetFillColor(kBlue);
-
-      summaryFirst->SetBarWidth(0.45);
-      summaryFirst->SetBarOffset(0.1);
-
-      summaryLast->SetBarWidth(0.4);
-      summaryLast->SetBarOffset(0.55);
-
-      summaryLast->SetMarkerSize(1.2);
-      summaryFirst->SetMarkerSize(1.2);
-
-      float max = (summaryFirst->GetMaximum() > summaryLast->GetMaximum()) ? summaryFirst->GetMaximum()
-                                                                           : summaryLast->GetMaximum();
-
-      summaryFirst->GetYaxis()->SetRangeUser(0., std::max(0., max * 1.40));
-
-      summaryFirst->Draw("b text0");
-      summaryLast->Draw("b text0 same");
-
-      TLegend legend = TLegend(0.52, 0.80, 0.98, 0.9);
-      legend.SetHeader(Form("#LT %s #GT value comparison", TypeName[myType]),
-                       "C");  // option "C" allows to center the header
-      legend.AddEntry(
-          summaryLast.get(),
-          ("IOV: #scale[1.2]{" + std::to_string(std::get<0>(lastiov)) + "} | #color[4]{" + std::get<1>(lastiov) + "}")
-              .c_str(),
-          "F");
-      legend.AddEntry(
-          summaryFirst.get(),
-          ("IOV: #scale[1.2]{" + std::to_string(std::get<0>(firstiov)) + "} | #color[2]{" + std::get<1>(firstiov) + "}")
-              .c_str(),
-          "F");
-      legend.SetTextSize(0.025);
-      legend.Draw("same");
-
-      std::string fileName(m_imageFileName);
-      canvas.SaveAs(fileName.c_str());
-      return true;
-    }
-
-    void fillThePerModuleMap(const std::shared_ptr<SiPixelGainCalibrationOffline>& payload,
-                             AvgMap& map,
-                             gainCalibPI::type theType) {
-      std::vector<uint32_t> detids;
-      payload->getDetIds(detids);
-
-      for (const auto& d : detids) {
-        auto range = payload->getRange(d);
-        int numberOfRowsToAverageOver = payload->getNumberOfRowsToAverageOver();
-        int ncols = payload->getNCols(d);
-        int nRocsInRow = (range.second - range.first) / ncols / numberOfRowsToAverageOver;
-        int nrows = payload->getNumberOfRowsToAverageOver() * nRocsInRow;
-
-        //int nrows = 80 * 2;  // rows in x
-        //int ncols = 52 * 8;  // cols in y
-
-        auto rangeAndCol = payload->getRangeAndNCols(d);
-        bool isDeadColumn;
-        bool isNoisyColumn;
-
-        float sumOfX(0.);
-        int nPixels(0);
-        for (int col_iter = 0; col_iter < ncols; col_iter++) {
-          for (int row_iter = 0; row_iter < nrows; row_iter++) {
-            nPixels++;
-            switch (theType) {
-              case gainCalibPI::t_gain:
-                sumOfX += payload->getGain(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              case gainCalibPI::t_pedestal:
-                sumOfX += payload->getPed(
-                    col_iter, row_iter, rangeAndCol.first, rangeAndCol.second, isDeadColumn, isNoisyColumn);
-                break;
-              default:
-                edm::LogError("SiPixelGainCalibrationOffline_PayloadInspector")
-                    << "Unrecognized type " << theType << std::endl;
-                break;
-            }  // switch on the type
-          }    // rows
-        }      // columns
-        // fill the return value map
-        map[d] = sumOfX / nPixels;
-      }  // loop on the detId
-    }
-  };
-
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineByRegionComparisonSingleTag
-      : public SiPixelGainCalibrationOfflineByRegionComparisonBase<myType> {
-  public:
-    SiPixelGainCalibrationOfflineByRegionComparisonSingleTag()
-        : SiPixelGainCalibrationOfflineByRegionComparisonBase<myType>() {
-      this->setSingleIov(false);
-    }
-  };
-
-  template <gainCalibPI::type myType>
-  class SiPixelGainCalibrationOfflineByRegionComparisonTwoTags
-      : public SiPixelGainCalibrationOfflineByRegionComparisonBase<myType> {
-  public:
-    SiPixelGainCalibrationOfflineByRegionComparisonTwoTags()
-        : SiPixelGainCalibrationOfflineByRegionComparisonBase<myType>() {
-      this->setTwoTags(true);
-    }
-  };
-
-  typedef SiPixelGainCalibrationOfflineByRegionComparisonSingleTag<gainCalibPI::t_gain>
-      SiPixelGainCalibOfflineGainByRegionComparisonSingleTag;
-  typedef SiPixelGainCalibrationOfflineByRegionComparisonSingleTag<gainCalibPI::t_pedestal>
-      SiPixelGainCalibOfflinePedestalByRegionComparisonSingleTag;
-
-  typedef SiPixelGainCalibrationOfflineByRegionComparisonTwoTags<gainCalibPI::t_gain>
-      SiPixelGainCalibOfflineGainByRegionComparisonTwoTags;
-  typedef SiPixelGainCalibrationOfflineByRegionComparisonTwoTags<gainCalibPI::t_pedestal>
-      SiPixelGainCalibOfflinePedestalByRegionComparisonTwoTags;
+  using SiPixelGainCalibrationOfflineGainsValues =
+      gainCalibHelper::SiPixelGainCalibrationValues<gainCalibHelper::gainCalibPI::t_gain, SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibrationOfflinePedestalsValues =
+      gainCalibHelper::SiPixelGainCalibrationValues<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                    SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibrationOfflineCorrelations =
+      gainCalibHelper::SiPixelGainCalibrationCorrelations<SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibrationOfflineGainsByPart =
+      gainCalibHelper::SiPixelGainCalibrationValuesByPart<gainCalibHelper::gainCalibPI::t_gain,
+                                                          SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibrationOfflinePedestalsByPart =
+      gainCalibHelper::SiPixelGainCalibrationValuesByPart<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                          SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainComparisonSingleTag =
+      gainCalibHelper::SiPixelGainCalibrationValueComparisonSingleTag<gainCalibHelper::gainCalibPI::t_gain,
+                                                                      SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalComparisonSingleTag =
+      gainCalibHelper::SiPixelGainCalibrationValueComparisonSingleTag<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                                      SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainComparisonTwoTags =
+      gainCalibHelper::SiPixelGainCalibrationValueComparisonTwoTags<gainCalibHelper::gainCalibPI::t_gain,
+                                                                    SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalComparisonTwoTags =
+      gainCalibHelper::SiPixelGainCalibrationValueComparisonTwoTags<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                                    SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainsBPIXMap =
+      gainCalibHelper::SiPixelGainCalibrationBPIXMap<gainCalibHelper::gainCalibPI::t_gain,
+                                                     SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalsBPIXMap =
+      gainCalibHelper::SiPixelGainCalibrationBPIXMap<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                     SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainsFPIXMap =
+      gainCalibHelper::SiPixelGainCalibrationFPIXMap<gainCalibHelper::gainCalibPI::t_gain,
+                                                     SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalsFPIXMap =
+      gainCalibHelper::SiPixelGainCalibrationFPIXMap<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                     SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainByRegionComparisonSingleTag =
+      gainCalibHelper::SiPixelGainCalibrationByRegionComparisonSingleTag<gainCalibHelper::gainCalibPI::t_gain,
+                                                                         SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalByRegionComparisonSingleTag =
+      gainCalibHelper::SiPixelGainCalibrationByRegionComparisonSingleTag<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                                         SiPixelGainCalibrationOffline>;
+
+  using SiPixelGainCalibOfflineGainByRegionComparisonTwoTags =
+      gainCalibHelper::SiPixelGainCalibrationByRegionComparisonTwoTags<gainCalibHelper::gainCalibPI::t_gain,
+                                                                       SiPixelGainCalibrationOffline>;
+  using SiPixelGainCalibOfflinePedestalByRegionComparisonTwoTags =
+      gainCalibHelper::SiPixelGainCalibrationByRegionComparisonTwoTags<gainCalibHelper::gainCalibPI::t_pedestal,
+                                                                       SiPixelGainCalibrationOffline>;
 
 }  // namespace
 
 PAYLOAD_INSPECTOR_MODULE(SiPixelGainCalibrationOffline) {
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationOfflineGainsValues);
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationOfflinePedestalsValues);
+  PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationOfflineCorrelations);
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationOfflineGainsByPart);
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationOfflinePedestalsByPart);
-  PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibrationCorrelations);
-  PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibOfflineGainComparisonSingleTag);
+  PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibOfflineGainComparisonSingleTag)
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibOfflinePedestalComparisonSingleTag);
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibOfflineGainComparisonTwoTags);
   PAYLOAD_INSPECTOR_CLASS(SiPixelGainCalibOfflinePedestalComparisonTwoTags);
